Changed timestamp from some arbitrary time from nanoTime to a string who's value is the current UTC date and time.
Changed functionality in the Queue so that duplicate usernames cannot be entered. (If "John Z" is already in queue, the name "John Z" cannot be added again.)

Changed functionality in the Queue so that the same user cannot join the Queue. Checks that webSocket is in the the server's map of sockets.

Changed functionality of "TA is ready to help" button. Pressing it removes the user from the server's map of sockets and usernames.

Fixed *Slight* Major Bug