Add files via upload
Added a StudentActor class, this keeps track of each user on the office hours website. Students cannot see the full queue but only the position they are in the queue. If a user hits the TA ready to help button they are registered as a TA and will be able to see the full queue.