Adding a user into DB. Using a class RegisterUser.
Next, will be distinguish between TA, Lecture Professor, and Students and every each of them has different page to look at it.

Secondly, we will have a board only for developer and the head of the website to control it.

Lastly, improve security and add more feature.

Will keep working on improving the idea during the winter break.

Android, IOS and Web apps ASAP.